fix: Fix collect_all type-coercion
